fixed issue with rating not showing correctly

+++ mozi-frontend/src/components/cards/MovieListCard.tsx
@@ -15,7 +15,7 @@ interface Props {
 
 export default function MovieListCard({movie}: Props) {
   const { t } = useTranslation();
-  const rating = parseInt(movie.rating)
+  const rating = parseFloat(movie.rating)
   return (
     <Link
       href={"/movie/" + movie.id}
